Fix (#30): Fix race-condition in parser-libraries while compiling
When using multiple threads while compiling, for example "make -j8" it
resulted in a race-condition. The parser-code was generated at the same
time when result was required by the code. Workaround until now was to
build with only one threads, which is very slow.
Now an additional prestep in the affected libraries was added to run
bison and flex in the correct order before the normal source-code.