Change (#76): indroduced new neuron-connections
As preparation for improving the GPU-performance, the backward-
connections between synapse-sections were replaced by a new
structure to hold ID's of all synapse-sections, which belongs
to a target-neuron-section. This way it will become possible
to let multiple GPU-threads work for the same neuron-section.
This will improve the speed especially for smaller networks.
The tradeoff is, that these sections have only a fixed limited
size, because buffering will be necessary for this approach,
but the GPUs have only a limited amount of local memory for this.