Fix: stupid size-validation in data-buffer
Within the data-buffer, when adding data to the buffer, there
was a check if the buffer has enough space for the data. This
check is stupid, because the function also contains a resize-
function after the check to increase the size of the buffer,
if the new data doesn't fit. This resize is never triggered
because of the validation. Because of this the validation
was removed.