Fix (#130): remove cluster at start
Cluster are in-memory, so they don't survive a restart
of the backend. It left behind a empty broken cluster-
corpse in the database behind without the cluster itself.
This was more or less solved for now by removing all
cluster from database while starting.
In the future the last available checkpoint should be
loaded.