Remove (#63): Remove ssl from backend-api
Remove ssl-encryption from the api of the
backend, because this should be outsourced
into an reverse-proxy like nginx for example.
Keeping the ssl-encryption beside the non-ssl
version in the backend makes it more complex
especially when it comes to replacing the
currently used http-library, which will happen
in the near future.