Change (#217): cleanup api-processing
Now error-messages on server-side are only printed,
in case of an internal server error and not for
each not-found, bad request or other non-critical
error.
Also cleaned the api, because the handling of the
responses were a total mess. This make the code
of the api cleaner and more robust.