Change (#46, #227): update file-upload and websockets
The internal handling of temporary files was changed.
Now these have its own database-table. Also with a simple
bit-buffer the progress of the upload can be tracked,
with will be later used for the ability to continue
uploads.
Also now when uploading data over the websocket, it
checks, if the user is allowed to access the resouce.
This check was also applied to the direct cluster-
connection.