Change (#213): big update of core-structure
The core-structure and -processing got a big re-design.
This new structure should increase the gpu-performance,
because it makes objects more independent and better
parallelizable. Also the recursive loop was removed,
which was the biggest pain-point for the gpu. It also
solves the problem of the number of synapse-blocks,
which had to be defined in the cluster-template. Now
there is only a single central buffer for these objects,
which is shared by all clusters. This will also make
checkpoints smaller and interaction with the gpu easier.
The cpu-code now runs nearly the same way, like the
new gpu-code will do, without big performance loss.
This new similarity will maintaining of the code much
easier.

At the moment only the cpu-part is updated and tested
for the new structure. Update of the gpu-kernel and fixing
checkpoints will be done in another commit to not blow
up the commit-size even more now.