Change (#251): update CUDA-kernel
Updated the structure of the CUDA-kernel in order
to massively increase the performance compared to
all previous versions. Now the gpu can share the
synapse-blocks for efficiently over the blocks of
the gpu.
Also fixed a bug, which resulted in a mismached
when processing the connection-blocks.
Added and updated comments in the CUDA-kernel.