Add (#230): re-added reduction-process
To limit the growth of the network, a reduction-process
was added, which removes synapses with too low activity-
rate and also removed empty synapse-sections again to
make them avaialbe for re-using them. This is still very
experimental, but already results in a very big reduction
of the used synapses. Also the	growth-rate over time
is much more limited now.
Only implementated for the cpu-path at the moment.