Fix (#257): fix memory-allocation
The cuda gpu's also allocating memory on the host based
in the amount of memory on the gpu. Because the host with
its cpu's is alloced first and doesn't know the memory on
the gpu, this could result in the allocation of more
memory, than avaialbe. This was fixed by initializing the
cuda gpu'S first.