Add (#192): add multi-threading
Added worker-thread, which do now the processing. There will be
as much worker-thread initialized, like cpu-threads exist. The
bricks are still processed one after the other, the bricks are
split blockwise, like in the gpu-case. These blocks are written
into a queue, where the worker-thread can take them. This way
multiple threads can work at the same time on the same brick
of the same cluster.