Change (#277): reduce cpu usage in idle state
In case the task-queue of the worker is empty, the worker
is not blocked and waked up, when a new task was added to
the queue. This blocking state is a delayed by a small
amount of time, to avoid race-conditions, when adding
tasks and blocking threads appears at the same time.