Fix (#280): fix measurements for virtual machines
In virtual machines, it can happen, that the package-ids
doesn't start with 0. This broke the measurement-loops,
which tried to iterate over the packages starting with 0.
This was fixed by not using the id anymore, but instead
the position in the buffer for access, to ensure that it
always starts at 0.