Change (#311): restructure cluster-data
Some of the cluster-data were still an old dirty binary-blob,
which remained from an older version, where all of the cluster-
data were combined into one single byte-buffer for faster
read and write of checkpoints. This is not the case anymore,
so the remaining parts of the binary were also replaced by
cpp-constructs for cleaner code and easier updates.