Change (#307, #323): rework checkpoints
Complete rework of the process of how-to read and
write checkpoints or in general, how-to serialize
and deserialize cluster. Because the checkpoints
required an update in context of the last brick-
changes, it was done together with the planned
checkpoint rework to avoid double workload. The
checkpoints are now only partially and not fully
cached anymore, which reduces the necessary amount
of memory exremely. Beside this the process is now
much more robust agains broken inputs.