Change (#337): rework dataset input and output
Now the datasets for train- and request-tasks are not
longer loaded completely into the memory, but instead
are read only in blocks from the disc, bases on the
rows requested by the current task.
The request-results were removed. Instead of this,
results of request-tasks are now written into a
dataset too and are now longer limited to only one
value per cycle.
Also each input-brick can now get data from a different
dataset and each output-brick can write results into
a different dataset.