Change (#337): update struct header
Made preallocated size of header for checkpoints
and datasets from 512 to 4096 byte to have more
room to add new fields without breaking too much.

Also used a new name-struct for a better generic
handling of name-structs, which are used for
serialization.

Changed name and id sizes from 256 to 254 bytes
for the new struct.