feat(#317): use connections for resize-info
The connection-objects are now the base for the
resize of the cluster. This removes the possibilty
of race-conditions in case that multiple connections
of the same neuron have to be split. On the other
hand it is the second part of the split of the
processing. This enables new possibilities for
in regard of gpu-processing and multi-node processing.