feat(#325): new attempt for gpu-support (incomplete)
Update the CUDA-kernel and connected functions in order to
re-enable the gpu-support again. Tests failed so far, because
of a very weird bug, which wasn't found so far. Seems like a
memory-corrumption somewhere, but when the new ASan memory-
check is enabled, CUDA doesn't work in general. To not waste
much more time at the moment for the debugging, the current
changes were added anyway to the develop-branch in order to
try it another time again. Until then the CUDA-support is
hard disabled in the source-code.