feat(#358): removed external auth endpoint
The external auth endpoint was removed. This endpoint
was originally added in the old micro-service
architecture. The the current state of the program
it doesn't make much sende the keep this as public
REST-API endpoint. Now a simple internal function
is used instead of the endpoint. This also reduces
the overhead of the token-validation.